added repositories functions in tution
